requirements.txt
requests==2.31.0
discord.py==2.2.3
deep-translator==1.11.1